add pages components, welcome view, login view, register view
