+++ lab-notes/src/App.js
@@ -1,18 +1,16 @@
 /* eslint-disable react/jsx-no-undef */
 /* eslint-disable react/react-in-jsx-scope */
-// import logo from './logo.svg';
-import './App.css';
+import "./App.css";
 import { Routes, Route } from "react-router-dom";
-import Welcome from './components/pages/welcome';
+import Welcome from "./components/pages/welcome";
 // import SignUp from './components/pages/register';
 // import SignIn from './components/pages/login';
-import RegisterUser from './components/pages/registerUser';
-import LoginUser from './components/pages/loginUser';
-import Home from './components/pages/home';
-import Homepage from './components/pages/homepage';
+import RegisterUser from "./components/pages/register/registerUser";
+import LoginUser from "./components/pages/login/loginUser";
+// import Home from './components/pages/home';
+import Homepage from "./components/pages/home/homepage";
 
-
-import { AuthProvider } from './context/authContext';
+import { AuthProvider } from "./context/authContext";
 
 function App() {
   return (
@@ -22,12 +20,12 @@ function App() {
           <Route path="/" element={<Welcome />} />
           <Route path="login" element={<LoginUser />} />
           <Route path="register" element={<RegisterUser />} />
-          <Route path="home" element={<Home />} />
+          {/* <Route path="home" element={<Home />} /> */}
           <Route path="homepage" element={<Homepage />} />
+          <Route path="*" element={<>NOT FOUND</>} />
         </Routes>
       </AuthProvider>
     </div>
-
   );
 }
