+++ lab-notes/src/components/pages/login/index.js
@@ -1,0 +1,2 @@
+//barrels?
+export { default as Login } from "./loginUser";
