+++ lab-notes/src/components/pages/register/registerUser.styles.js
@@ -1,0 +1,6 @@
+export const firstBox =  {
+    display: "flex",
+    justifyContent: "center",
+    flexDirection: "column",
+    height: "100vh"
+}
